requirements.txt
django
python-dotenv
mysqlclient
cloudinary
django-cloudinary-storage
gunicorn
whitenoise
qrcode
Pillow
